// src/types/qc-check.ts
export type QcCheckType = {
  id: number
  srNo: number
  testName: string
  instrumentName: string
  level1: string
  level2: string
  level3: string
  doneOn: string
  doneBy: string
  profile: string
  date: string
  sampleId: string
  referenceId: string
  genderName: string
  parameter: string
}